Componentalized the modal, am  I using the right words for these?
